almost fixed offer tests
