+++ testcafe/helpers/navigations.js
@@ -26,7 +26,9 @@ export const navigateToOffererAs = (user, offerer) => async t => {
   const searchInput = Selector('#search')
   const submitButton = Selector('button[type="submit"]')
 
-  const offererAnchor = Selector("a[href^='/structures/']").nth(0)
+  const offererAnchor = Selector('li.offerer-item')
+    .nth(0)
+    .find("a[href^='/structures/']")
 
   await navigateToOfferersAs(user)(t)
 
@@ -112,7 +114,9 @@ export const navigateToNewOfferAs = (user, offerer, venue) => async t => {
 export const navigateToOfferAs = (user, offer) => async t => {
   const searchInput = Selector('#search')
   const submitButton = Selector('button[type="submit"]')
-  const offerAnchor = Selector('.name').nth(0)
+  const offerAnchor = Selector('li.offer-item')
+    .nth(0)
+    .find('.name')
 
   await t.useRole(createUserRole(user))
