+++ testcafe/helpers/offerers.js
@@ -32,7 +32,8 @@ export const OFFERER_WITH_PHYSICAL_VENUE_WITH_NO_IBAN = {
 export const OFFERER_WITH_PHYSICAL_VENUE_WITH_IBAN = {
   address: 'RUE DIDEROT',
   city: 'Aulnay-sous-Bois',
-  keywordsString: 'Danse Jazz Dans Tes Bottes',
+  // TODO: Danse Jazz Dans Tes Bottes does not work
+  keywordsString: 'Danse Jazz Bottes',
   name: 'Danse Jazz Dans Tes Bottes',
   postalCode: '93600',
   siren: '222222229',
