+++ testcafe/helpers/offers.js
@@ -1,4 +1,36 @@
-export const OFFER_WITH_NO_OFFERER_IBAN_WITH_NO_VENUE_IBAN = {
+export const FUTURE_EVENT_OFFER_WITH_OFFERER_IBAN_WITH_NO_VENUE_IBAN = {
+  eventDescription: [
+    'Alors que les licenciements sont devenus des plans de sauvegarde de l’emploi, ',
+    'ou que votre banquier se veut votre partenaire, ',
+    'il est temps de se poser certaines questions. ',
+    'Avec d’autres travailleurs socioculturels, ',
+    'lassés des euphémismes et des mensonges du langage du pouvoir, ',
+    'Franck Lepage s’est lancé dans cette bataille très politique : celle des mots. ',
+    "Atelier d'initiation pour reconnaître tout ce qui est du pipotron dans vos lectures de tous les jours. ",
+    "Suivi d'une séance de dédicaces.",
+  ].join(''),
+  eventDurationMinutes: '120',
+  eventName: 'Rencontre avec Franck Lepage',
+  eventType: 'Conférence — Débat — Dédicace',
+}
+
+export const FUTURE_MUSIC_EVENT_OFFER_WITH_OFFERER_IBAN_WITH_NO_VENUE_IBAN = {
+  eventDescription: 'Venez re découvrir PNL en accoustique, sans auto-tune',
+  eventDurationMinutes: '90',
+  eventName: 'Concert de PNL Unplugged',
+  eventType: 'Musique (Concerts, Festivals)',
+  eventMusicType: 'Hip-Hop/Rap',
+  eventMusicSubType: 'Rap Alternatif',
+}
+
+export const FUTURE_VIRTUAL_THING_OFFER_WITH_OFFERER_IBAN_WITH_NO_VENUE_IBAN = {
+  thingDescription: 'Jeux vidéo de test',
+  thingName: 'Jeux vidéo abonnement de test',
+  thingType: 'Jeux Vidéo',
+  thingUrl: 'http://www.example.com',
+}
+
+export const EVENT_OFFER_WITH_OFFERER_IBAN_WITH_NO_VENUE_IBAN = {
   // city: "Drancy",
   // TODO does not work yet with "C'est"
   keywordsString: 'notre prooooojecteur Dorothy',
